Store rendered text using GL_LUMINANCE_ALPHA instead of GL_RGBA.
We need to pack out the rows so each row starts on a 32-bit boundary.
No idea why we need to do this. It's probably mentioned in some
obscure corner of the manual, or it might be an OSX OpenGL bug.

--HG--
extra : convert_revision : svn%3A14d46d22-621c-0410-bb3d-6f67920f7d95/trunk%4096